code edits

+++ basic/basic04x_count_char.c
@@ -2,12 +2,12 @@
 
 int main(){
 
-	double nc;
+	int nc;
 
 	for(nc=0; getchar()!= EOF; ++nc)
 		;
 
-	printf("%.0f\n",nc);
+	printf("%d\n",nc);
 
 	return 0;
 }
